CareConnect-Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-adk
google-cloud-aiplatform
google-genai